# api/docs/swagger.yaml
basePath: /v1
definitions:
  customer.Address:
    properties:
      District:
        type: string
      Street:
        type: string
    type: object
  customer.AddressResponse:
    properties:
      District:
        type: string
      Street:
        type: string
      id:
        type: integer
    type: object
  customer.CustomerFullInfo:
    properties:
      Adderesses:
        items:
          $ref: '#/definitions/customer.AddressResponse'
        type: array
      Bio:
        type: string
      CreatedAt:
        type: string
      DeletedAt:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      Id:
        type: string
      LastName:
        type: string
      PhoneNumber:
        type: string
      Posts:
        items:
          $ref: '#/definitions/customer.PostResponse'
        type: array
      Reviews:
        items:
          $ref: '#/definitions/customer.ReviewList'
        type: array
      UpdatedAt:
        type: string
    type: object
  customer.CustomerRequest:
    properties:
      Adderesses:
        items:
          $ref: '#/definitions/customer.Address'
        type: array
      Bio:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      Id:
        type: string
      LastName:
        type: string
      PassWord:
        type: string
      PhoneNumber:
        type: string
      RefreshToken:
        type: string
      UserName:
        type: string
    type: object
  customer.CustomerResponse:
    properties:
      Adderesses:
        items:
          $ref: '#/definitions/customer.AddressResponse'
        type: array
      Bio:
        type: string
      CreatedAt:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      Id:
        type: string
      LastName:
        type: string
      PhoneNumber:
        type: string
      UpdatedAt:
        type: string
    type: object
  customer.CustomerResponsePost:
    properties:
      Adderesses:
        items:
          $ref: '#/definitions/customer.AddressResponse'
        type: array
      Bio:
        type: string
      CreatedAt:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      Id:
        type: string
      LastName:
        type: string
      PhoneNumber:
        type: string
      Post:
        items:
          $ref: '#/definitions/customer.PostResponse'
        type: array
      UpdatedAt:
        type: string
    type: object
  customer.CustomerUpdate:
    properties:
      Adderesses:
        items:
          $ref: '#/definitions/customer.AddressResponse'
        type: array
      Bio:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      Id:
        type: string
      LastName:
        type: string
      PhoneNumber:
        type: string
    type: object
  customer.ListCustomers:
    properties:
      ActiveCustomers:
        items:
          $ref: '#/definitions/customer.CustomerFullInfo'
        type: array
      DeletedCustomers:
        items:
          $ref: '#/definitions/customer.CustomerFullInfo'
        type: array
    type: object
...
  customer.MediasResponse:
    properties:
      Id:
        type: integer
      Link:
        type: string
      Name:
        type: string
      Type:
        type: string
    type: object
  customer.PostResponse:
    properties:
      CreatedAt:
        type: string
      Description:
        type: string
      Id:
        type: integer
      Media:
        items:
          $ref: '#/definitions/customer.MediasResponse'
        type: array
      Name:
        type: string
      UpdatedAt:
        type: string
    type: object
  customer.ReviewList:
    properties:
      description:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      review:
        type: number
    type: object
...
  post.AddressResponse:
    properties:
      District:
        type: string
      Street:
        type: string
      id:
        type: integer
    type: object
  post.ListAllPostResponse:
    properties:
      ActivePost:
        items:
          $ref: '#/definitions/post.PostReviewResponse'
        type: array
      DeletedPost:
        items:
          $ref: '#/definitions/post.PostReviewResponse'
        type: array
    type: object
  post.MediasRequest:
    properties:
      Link:
        type: string
      Name:
        type: string
      Type:
        type: string
    type: object
  post.MediasResponse:
    properties:
      Link:
        type: string
      Name:
        type: string
      Type:
        type: string
      id:
        type: integer
    type: object
  post.PostRequest:
    properties:
      CustomerId:
        type: string
      Description:
        type: string
      Media:
        items:
          $ref: '#/definitions/post.MediasRequest'
        type: array
      Name:
        type: string
    type: object
  post.PostResponse:
    properties:
      CreatedAt:
        type: string
      CustomerId:
        type: string
      Description:
        type: string
      Id:
        type: integer
      Media:
        items:
          $ref: '#/definitions/post.MediasResponse'
        type: array
      Name:
        type: string
      UpdatedAt:
        type: string
    type: object
  post.PostResponseCustomer:
    properties:
      Bio:
        type: string
      CreatedAt:
        type: string
      CustomerId:
        type: string
      Description:
        type: string
      FirstName:
        type: string
      Id:
        type: integer
      LastName:
        type: string
      Media:
        items:
          $ref: '#/definitions/post.MediasResponse'
        type: array
      Name:
        type: string
      UpdatedAt:
        type: string
      adderesses:
        items:
          $ref: '#/definitions/post.AddressResponse'
        type: array
      email:
        type: string
      phonenumber:
        type: string
      reviews:
        items:
          $ref: '#/definitions/post.ReviewRespList'
        type: array
    type: object
  post.PostReviewResponse:
    properties:
      Count:
        type: integer
      CreatedAt:
        type: string
      CustomerId:
        type: string
      DeletedAt:
        type: string
      Description:
        type: string
      Id:
        type: integer
      Media:
        items:
          $ref: '#/definitions/post.MediasResponse'
        type: array
      Name:
        type: string
      Overall:
        type: number
      UpdatedAt:
        type: string
      reviews:
        items:
          $ref: '#/definitions/post.ReviewRespList'
        type: array
    type: object
  post.PostUpdate:
    properties:
      CustomerId:
        type: string
      Description:
        type: string
      Id:
        type: integer
      Media:
        items:
          $ref: '#/definitions/post.MediasResponse'
        type: array
      Name:
        type: string
    type: object
  post.ReviewRespList:
    properties:
      customer_id:
        type: string
      description:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      review:
        type: number
    type: object
...
host: localhost:8070
info:
  contact:
    email: asliddinvstalim@gmail.com
    name: Asliddin
    url: https://t.me/asliddindeh
  description: This is reivew api
  termsOfService: not much usefull
  title: Review api
  version: "1.0"
paths:
...
  /customer:
    post:
      consumes:
      - application/json
      description: this func create customer with
      parameters:
      - description: Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customer.CustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
      security:
      - BearerAuth: []
      summary: create customer with info
      tags:
      - customer
  /customer/{id}:
    get:
      consumes:
      - application/json
      description: this func get customer info
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
      security:
      - BearerAuth: []
      summary: get customer info
      tags:
      - customer
  /customer/delete/{id}:
    delete:
      consumes:
      - application/json
      description: this func delete customer
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
      security:
      - BearerAuth: []
      summary: delete customer
      tags:
      - customer
  /customer/list:
    get:
      consumes:
      - application/json
      description: this func get all customers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.ListCustomers'
      security:
      - BearerAuth: []
      summary: get all customers
      tags:
      - customer
  /customer/post/{id}:
    get:
      consumes:
      - application/json
      description: this func get customer with post
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponsePost'
      security:
      - BearerAuth: []
      summary: get customer with post
      tags:
      - customer
  /customer/update:
    patch:
      consumes:
      - application/json
      description: this func update customers
      parameters:
      - description: Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customer.CustomerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.CustomerResponse'
      security:
      - BearerAuth: []
      summary: update customers
      tags:
      - customer
...
  /post:
    post:
      consumes:
      - application/json
      description: this func create post
      parameters:
      - description: Post
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/post.PostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/post.PostResponse'
      security:
      - BearerAuth: []
      summary: create post with info
      tags:
      - post
  /post/{id}:
    get:
      consumes:
      - application/json
      description: this func get post
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResponseCustomer'
      security:
      - BearerAuth: []
      summary: get post
      tags:
      - post
  /post/delete/{id}:
    delete:
      consumes:
      - application/json
      description: this func delete post
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
      security:
      - BearerAuth: []
      summary: delete post
      tags:
      - post
  /post/list:
    get:
      consumes:
      - application/json
      description: this func get posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.ListAllPostResponse'
      security:
      - BearerAuth: []
      summary: get posts
      tags:
      - post
  /post/search/{page}/{limit}/{parameters}/{orderby}:
    get:
      consumes:
      - application/json
      description: this func search post
      parameters:
      - description: page
        in: path
        name: page
        required: true
        type: integer
      - description: limit
        in: path
        name: limit
        required: true
        type: integer
      - description: parameters
        in: path
        items:
          type: string
        name: parameters
        required: true
        type: array
      - description: orderby
        in: path
        name: orderby
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
      security:
      - BearerAuth: []
      summary: search post
      tags:
      - post
  /post/update:
    patch:
      consumes:
      - application/json
      description: this func update post
      parameters:
      - description: Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/post.PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResponse'
      security:
      - BearerAuth: []
      summary: update post
      tags:
      - post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"